Cambio Observador

+++ src/Observador/EstadoBusqueda.java
@@ -5,13 +5,13 @@ package Observador;
  * @author Vicente
  */
 public class EstadoBusqueda extends Observable{
-    private float maximo;
+    private double maximo;
     
     public EstadoBusqueda(){
         maximo= Float.NEGATIVE_INFINITY;
     }
     
-    public synchronized void setMaximo(float maxBuscador){
+    public synchronized void setMaximo(double maxBuscador){
         if(maximo < maxBuscador){
             maximo= maxBuscador;
             notificar(maximo);
